PIXEL_PLAY/requirements.txt
streamlit
opencv-python
mediapipe
pillow
numpy
pygame
streamlit-custom-page